+++ apps/admin/app/containers/BrokerCAD/VerifyAccount/index.tsx
@@ -4,7 +4,7 @@ import { Form, Header, Button, List, Message } from 'semantic-ui-react';
 import { UxAddress } from '@thecointech/shared/components/UxAddress';
 import messages from './messages'
 import { getShortCode, NormalizeAddress } from '@thecointech/utilities';
-import { TheSigner } from '@thecointech/shared/SignerIdent';
+import { TheSigner } from '@thecointech/utilities/SignerIdent';
 import { setUserVerified } from '@thecointech/broker-db/user';
 import { createReferrer, getReferrersCollection } from '@thecointech/broker-db/referrals';
 import { DateTime } from 'luxon';
