+++ libs/shared/src/utils/storageSync.ts
@@ -1,6 +1,6 @@
 import { AccountState, DefaultAccountValues } from '../containers/Account/types';
 import { AccountMap } from '../containers/AccountMap';
-import { isSigner, SignerIdent } from '../SignerIdent';
+import { isSigner, SignerIdent } from '@thecointech/utilities/SignerIdent';
 import { IsValidAddress, NormalizeAddress } from '@thecointech/utilities';
 import { Deprecated_GetStored } from './storageSync_deprecated';
